11: Server - Refactor - extract shared worker utilies and reuse in scheduler and dispatcher
