+++ server/src/workers/scheduler.worker.ts
@@ -1,17 +1,9 @@
 import Schedule from '../models/schedule.js';
 
 import { CronExpressionParser } from 'cron-parser';
-import type {
-    SchedulerTickHandler,
-    SchedulerLogger,
-    SchedulerSignal,
-    SchedulerProcessLike,
-    SchedulerWorkerOptions,
-    SchedulerTickContext,
-    SchedulerTickResult,
-    SchedulerTickDueSchedule,
-} from '../models/types.js';
+import type { SchedulerTickHandler, SchedulerLogger, SchedulerSignal, SchedulerProcessLike, SchedulerWorkerOptions, SchedulerTickContext, SchedulerTickResult, SchedulerTickDueSchedule } from '../models/types.js';
 import { ScheduledJobService } from '../services/scheduledJobService.js';
+import { normalizeBoolean, normalizePositiveInteger, resolveProcess, hasUnref } from './worker.utils.js';
 
 const DEFAULT_POLL_INTERVAL_MS = 30_000;
 const LOG_PREFIX = '[SchedulerWorker]';
@@ -91,78 +83,6 @@ const defaultTickHandler: SchedulerTickHandler = async (
     };
 };
 
-function normalizeBoolean(
-    value: string | undefined,
-    fallback: boolean
-): boolean {
-    if (value === undefined) {
-        return fallback;
-    }
-
-    const normalized = value.trim().toLowerCase();
-
-    if (['1', 'true', 'yes', 'on'].includes(normalized)) {
-        return true;
-    }
-
-    if (['0', 'false', 'no', 'off'].includes(normalized)) {
-        return false;
-    }
-
-    return fallback;
-}
-
-function normalizeInterval(value: string | undefined, fallback: number): number {
-    if (value === undefined) {
-        return fallback;
-    }
-
-    const parsed = Number.parseInt(value, 10);
-
-    if (Number.isNaN(parsed) || parsed <= 0) {
-        return fallback;
-    }
-
-    return parsed;
-}
-
-function resolveProcess(
-    customProcess?: SchedulerProcessLike
-): SchedulerProcessLike {
-    if (customProcess) {
-        return customProcess;
-    }
-
-    const candidate = (globalThis as { process?: unknown }).process as
-        | SchedulerProcessLike
-        | undefined;
-
-    if (
-        candidate &&
-        typeof candidate.on === 'function' &&
-        candidate.env &&
-        typeof candidate.exit === 'function'
-    ) {
-        return candidate;
-    }
-
-    return {
-        env: {},
-        on: () => undefined,
-        off: () => undefined,
-        exit: () => undefined,
-        kill: () => undefined,
-    };
-}
-
-function hasUnref(timer: unknown): timer is { unref: () => void } {
-    return (
-        typeof timer === 'object' &&
-        timer !== null &&
-        typeof (timer as { unref?: unknown }).unref === 'function'
-    );
-}
-
 export class SchedulerWorker {
     private readonly enabled: boolean;
     private readonly intervalMs: number;
@@ -189,7 +109,7 @@ export class SchedulerWorker {
 
         this.intervalMs =
             options.intervalMs ??
-            normalizeInterval(
+            normalizePositiveInteger(
                 this.proc.env.SCHEDULER_POLL_INTERVAL_MS,
                 DEFAULT_POLL_INTERVAL_MS
             );
